fix: moved PolicyPostDTO
